requirements.txt
pyotp==2.4.1
pyxdg==0.26
